+++ examples/shred_pymarc/shred.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+set -euo pipefail
+bash2gitlab shred --in-file .gitlab-ci.yml --out src