+++ test/test_tui.py
@@ -1,0 +1,5 @@
+import bash2gitlab.tui
+
+
+def test_imports():
+    assert dir(bash2gitlab.tui)